Update README.md
In this exercise I built a data pipeline that extracts data from MongoDB in the cloud and in the IBGE API, depositing it in an AWS Data Lake (S3). After ingesting data in S3, I made the processed and filtered data available only to the public of interest. With the data in the database, I did some BI queries.
In the other folder I made the same pipeline but in Databricks .